FEAT [TOOLS] [logic] [no-tools-called-bypass] When no tools are called - it doesn't trigger an error anymore, instead it wraps last message into an attempt_completion call.
Done to:

- fix scenarios when model re-generates same message
but just to include it in attempt_completion call,
which wastes money for no reason.

- fix scenarios when model responds correctly in plain
text, but task progression fails in an error because
model can't do that inside a task (which is not right, imho)

- fix subtask calling mechanism that will broken because
of current edits, if model won't invoke attempt completion,
subtask won't be considered as completed and its output
won't be passed to the parent task.