FEAT [UI] [tasks-view] [show-more-tasks] Shows more tasks + makes some small changes to make everything look a bit compact and better
removes font-light from tasks text because it
renders weirdly.

Also because of weird logic of filtering tasks,
that only applies to paged tasks in history view,
I had to set pretty high number for that view
so on recent tasks it will filter last 10 without
parent more often, instead of like showing
8 tasks when in reality there's like a 30 without
a parent, but just on next page.